👌 Move resourcetree package to controller/common/resourcetree
