+++ pkg/controller/application/reconciliation/container_repository_reconciler.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/oslokommune/okctl/pkg/client"
 	clientCore "github.com/oslokommune/okctl/pkg/client/core"
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 )
 
 // containerRepositoryReconciler contains service and metadata for the relevant resource
