+++ pkg/controller/cluster/reconciliation/argocd_reconciler.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/miekg/dns"
 	"github.com/oslokommune/okctl/pkg/client"
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 )
 
 // NodeType returns the relevant ResourceNodeType for this reconciler
