+++ pkg/controller/cluster/reconciliation/awsloadbalancercontroller_reconciler.go
@@ -8,7 +8,7 @@ import (
 	"github.com/oslokommune/okctl/pkg/controller/common/reconciliation"
 
 	"github.com/oslokommune/okctl/pkg/client"
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 )
 
 // NodeType returns the relevant ResourceNodeType for this reconciler
