+++ pkg/controller/cluster/reconciliation/cleanup_sg_reconciler.go
@@ -12,7 +12,7 @@ import (
 
 	clientCore "github.com/oslokommune/okctl/pkg/client/core"
 
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 )
 
 type cleanupSGReconciler struct {
