+++ pkg/controller/cluster/reconciliation/cluster_reconciler.go
@@ -11,7 +11,7 @@ import (
 	"github.com/oslokommune/okctl/pkg/config/constant"
 
 	"github.com/oslokommune/okctl/pkg/client"
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 )
 
 // clusterReconciler contains service and metadata for the relevant resource
