+++ pkg/controller/cluster/reconciliation/postgres_group_reconciler.go
@@ -3,7 +3,7 @@ package reconciliation
 import (
 	clientCore "github.com/oslokommune/okctl/pkg/client/core"
 	"github.com/oslokommune/okctl/pkg/controller/common/reconciliation"
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 )
 
 // PostgresGroupReconciler handles reconciliation for dummy nodes (e.g. the root node) and acts as a template for other
