+++ pkg/controller/cluster/reconciliation/sevicequota_reconciler.go
@@ -9,7 +9,7 @@ import (
 	clientCore "github.com/oslokommune/okctl/pkg/client/core"
 	"github.com/oslokommune/okctl/pkg/config/constant"
 	"github.com/oslokommune/okctl/pkg/controller/common/reconciliation"
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 	"github.com/oslokommune/okctl/pkg/servicequota"
 	"github.com/pkg/errors"
 )
