+++ pkg/controller/common/reconciliation/reconciler.go
@@ -5,7 +5,7 @@ import (
 
 	clientCore "github.com/oslokommune/okctl/pkg/client/core"
 
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 )
 
 // Result contains information about the result of a Reconcile() call
