+++ pkg/controller/convert.go
@@ -20,7 +20,7 @@ import (
 
 	clientCore "github.com/oslokommune/okctl/pkg/client/core"
 
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 )
 
 // ExistingResources contains information about what services already exists in a cluster
