+++ pkg/controller/synchronize_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/oslokommune/okctl/pkg/config/constant"
 
-	"github.com/oslokommune/okctl/pkg/controller/resourcetree"
+	"github.com/oslokommune/okctl/pkg/controller/common/resourcetree"
 	"github.com/stretchr/testify/assert"
 )
